refactor: Number Class Computer와 User의 Number로 각각 분할
