+++ src/main/java/baseball/ComputerNumber.java
@@ -1,2 +1,22 @@
-package baseball;public class ComputerNumber {
+package baseball;
+
+import java.util.ArrayList;
+
+import static baseball.GameManager.*;
+import static camp.nextstep.edu.missionutils.Randoms.pickNumberInRange;
+
+public class ComputerNumber {
+
+    ArrayList<Integer> getComputerNumbers(){
+        ArrayList<Integer> computerNumbers = new ArrayList<>();
+
+        while(computerNumbers.size() < NUMBER_LENGTH){
+            int num = pickNumberInRange(startRange, endRange);
+            if(!computerNumbers.contains(num)){
+                computerNumbers.add(num);
+            }
+        }
+
+        return computerNumbers;
+    }
 }
